fixed overflow issue with millisecond field of sample_t struct

+++ firmware/include/collate.h
@@ -21,7 +21,7 @@ typedef struct {
     uint8_t hour;
     uint8_t minute;
     uint8_t second;
-    uint8_t millisecond;
+    uint16_t millisecond;
     bool valid_time;
 } sample_t;
